+++ libavfilter/vf_huesaturation.c
@@ -105,7 +105,7 @@ static int do_slice_##name##_##xall(AVFilterContext *ctx,            \
     const int process_h = frame->height;                             \
     const int slice_start = (process_h *  jobnr   ) / nb_jobs;       \
     const int slice_end   = (process_h * (jobnr+1)) / nb_jobs;       \
-    const int linesize = frame->linesize[0] / sizeof(type);          \
+    const ptrdiff_t linesize = frame->linesize[0] / sizeof(type);    \
     type *row = (type *)frame->data[0] + linesize * slice_start;     \
     const uint8_t offset_r = s->rgba_map[R];                         \
     const uint8_t offset_g = s->rgba_map[G];                         \
